#1431 Address review: dedupe MDT clone-index hack, share tooltip cleanup, fix handlebars translation convention
- Extracted the duplicated MDT clone-index hack (Siege of Boralus/Tol Dagor/Mists of Tirna
  Scithe) out of RaidMarkerImporter and PullImporter into a shared
  App\Service\MDT\Import\Traits\AppliesMdtCloneIndexHack trait.
- Moved the stray-tooltip-bubble cleanup out of enemyvisual.js into a reusable
  removeStrayTooltips() helper in util.js.
- enemy_tooltip_template's right_click_to_open_details/shift_right_click_to_assign_raid_marker
  translation keys now follow the site's _label suffix convention and are supplied automatically
  via getHandlebarsDefaultVariables() instead of being passed in manually.
