+++ resources/assets/js/handlebars/enemy_tooltip_template.handlebars
@@ -24,13 +24,13 @@
     {{/each}}
     <div class="row g-0 mt-1">
         <div class="col-12 italic">
-            {{right_click_to_open_details}}
+            {{right_click_to_open_details_label}}
         </div>
     </div>
     {{#if show_raid_marker_shortcut}}
         <div class="row g-0">
             <div class="col-12 italic">
-                {{shift_right_click_to_assign_raid_marker}}
+                {{shift_right_click_to_assign_raid_marker_label}}
             </div>
         </div>
     {{/if}}
